Documentation of code and user Avatar feature added

+++ src/features/comment/Model/comment.repository.js
@@ -1,3 +1,4 @@
+// Comments repository file is here all database activities is happening here.
 // Imports
 import { ObjectId } from "mongodb";
 import handleDatabaseError from "../../../errors/databaseError.js";
@@ -16,7 +17,10 @@ export default class CommentRepository{
             {
                 throw new  ApplicationError("No post found on this id", 404);
             }
-            const comments = await CommentModel.find({post: new ObjectId(postId)});
+            const comments = await CommentModel.find({post: new ObjectId(postId)}).populate({
+                path: 'user',
+                select: 'name email _id' // Specify the fields to include
+            });
             if(!comments || comments.length == 0)
             {
                 throw new  ApplicationError("No comments found on this post.", 404);
@@ -41,8 +45,11 @@ export default class CommentRepository{
                 post: new ObjectId(postId),
                 content: content,
             });
-            const savedComment = newComment.save();
-            return savedComment;
+            const savedComment = await newComment.save();
+            // Adding to post comments Array
+            post.comments.push(savedComment);
+            await post.save();
+            return savedComment.populate({path: 'user', select: 'name email _id'});
         } catch (error) {
             handleDatabaseError(error);
         }
@@ -60,6 +67,16 @@ export default class CommentRepository{
             if(String(comment.user) !== userID)
             {
                 throw new  ApplicationError("You are not allowed to delete this comment.", 404);
+            }
+            const post = await PostModel.findById(comment.post);
+            if (!post) {
+                throw new ApplicationError("Associated post not found.", 404);
+            }
+            // Remove comment ID from post's comments array
+            const index = post.comments.indexOf(commentId);
+            if (index > -1) {
+                post.comments.splice(index, 1);
+                await post.save();
             }
             const result = await CommentModel.findByIdAndDelete(commentId);
             return result;
@@ -81,7 +98,10 @@ export default class CommentRepository{
             }
             commentToUpdate.content = updatedContent; // Assuming 'content' is the field to be updated
             const updatedComment = await commentToUpdate.save();
-            return updatedComment;
+            return updatedComment.populate({
+                path: 'user',
+                select: 'name email _id' // Specify the fields to include
+            });
         } catch (error) {
             handleDatabaseError(error);
         }
